Gets DI working to inject at least at a basic level to access genvals, and cleans up some interfaces

+++ TaskQueueProcessor/src/NIST.CVP.TaskQueueProcessor/Providers/IDbProvider.cs
@@ -3,7 +3,7 @@ namespace NIST.CVP.TaskQueueProcessor.Providers
     public interface IDbProvider
     {
         ITask GetNextTask();
-        void DeleteCompletedTask(int taskId);
+        void DeleteCompletedTask(long taskId);
         void MarkTasksForRestart();
     }
 }
